simplify and update implementation of HasDamagingItem
This was an earlier revision of this function that Lunar suggested I use instead. (Although also correcting the `object` in the param should be `List<string>`)
Thank you Lunar for checking my work.

Notes: I find that calling a new list object every single time is clunky. I left a note about that in the code if anyone comes through in the future.