add a null safety net to HasDamagingItem
Hmm. Not sure how this could be an issue. If exclusions is empty, then the line `!exclusions.Contains()` should return "true" so that the item can be checked for.

In my prototyping, this worked as well.

Regardless, I am bug chasing, and bug leads to new code, so here's a safety net.