create CanSurviveBonkDamage
This was the most "rewriting" the logic I have done, and it was not even rewriting the logic per-se.

This function is a quick function to ensure that bonk damage can be survived.

I added this in part because repetition is getting cut out and packed into functions, and in part because I keep seeing this chunk and struggling to read it.

- fixed a typo on line 1617 (should fix that summary!)