move hidden skill levels into their own file
Some of this abstraction might be worse; I want to try something here. Sorry.

Basically: I am trying to remove that "Magic Number" that needs to be known? I am abstracting this into its own (subclass of helperfuncs), anyway, which means that I can have private things for only the class to see.
My hope is that this will help with the "argh, what's the actual name of that skill" or "argh, what level is this skill, again?"