Update HelperFunctions.cs
So I had this function idea in my previous code; I have been ignoring most of that code because... mainly because of the fact that that was 6 months ago and the randomizer has changed, significantly.

https://www.geeksforgeeks.org/c-sharp/c-sharp-params/

I looked over at my code and realized two things. 1. the naming was wrong (see the last commit). 2. *I left a love letter for future self... and then completely ignored said love letter for future self.*

So enjoy this lovely little function that is not *technically* better, but may be more readable. (also I need to fix `extraItems` now -.-;)