implement niche logic utils
Pivoting away from a file which defines all of the logic yada yada, instead focusing on utils. I might pivot back, but not for glitchless. We will see.

Starting out will be the 2 big hits: Niche and DifficultCombat logic. utils. Starting out with Niche which was driving me the most insane.