finish out the first pass of refactoring by simplifying glitched
So much to do now, but I can publish the branch finally.

I need to do this in order to be able to pull the changes incoming from Lunar working on the Ordon freestanding rupees, anyway.

HEY IF YOU SEE THIS BE WARNED UNTESTED CODE. Lol (not like the changes that are made will affect the randomizer because this isn't hooked to anything.)