write notes of where I see things used
I want to try to simplify things so I want to know where things are used...

Except I think I need to be focusing on LogicTokenizer most... which means I need to then parse *all of the logic files* to figure out what functions are used. And also get an understanding of LT.
right.

well the next commit is gonna be me focusing down on `GeneratePortalRooms` because I think I can clean it up...